requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.12.0 ; python_version >= "3.11" and python_version < "4.0"
asn1crypto==1.5.1 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.11" and python_version < "4"
boto3==1.42.4 ; python_version >= "3.11" and python_version < "4.0"
botocore==1.42.4 ; python_version >= "3.11" and python_version < "4.0"
cachetools==6.2.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.11.12 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and platform_python_implementation != "PyPy" and python_version < "4.0"
charset-normalizer==3.4.4 ; python_version >= "3.11" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.11" and python_version < "4"
cloudevents==1.12.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0"
cryptography==46.0.3 ; python_version >= "3.11" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.8.0 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.19.1 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
flask==3.1.2 ; python_version >= "3.11" and python_version < "4"
functions-framework==3.10.0 ; python_version >= "3.11" and python_version < "4"
google-api-core==2.28.1 ; python_version >= "3.11" and python_version < "4.0"
google-api-core[grpc]==2.28.1 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.43.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-appengine-logging==1.7.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-audit-log==0.4.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.5.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-logging==3.12.1 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-trace==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.72.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos[grpc]==1.72.0 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.14.3 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.76.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.76.0 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.11" and python_version < "4"
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.11 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==8.7.0 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
jmespath==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.3 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-api==1.34.1 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-exporter-gcp-trace==1.11.0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-instrumentation==0.55b1 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-resourcedetector-gcp==1.11.0a0 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-sdk==1.34.1 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-semantic-conventions==0.55b1 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
pg8000==1.31.4 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.11" and python_version < "4.0"
protobuf==6.33.2 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.10 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.11" and platform_python_implementation != "PyPy" and python_version < "4.0" and implementation_name != "PyPy"
pydantic-core==2.41.5 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.11" and python_version < "4.0"
pymysql==1.1.2 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.2.1 ; python_version >= "3.11" and python_version < "4.0"
python-jose==3.5.0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.5 ; python_version >= "3.11" and python_version < "4.0"
responses==0.25.8 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.11" and python_version < "4"
s3transfer==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
scramp==1.4.6 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.43 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.50.0 ; python_version >= "3.11" and python_version < "4"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.6.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn-worker==0.4.0 ; python_version >= "3.11" and python_version < "4"
uvicorn==0.38.0 ; python_version >= "3.11" and python_version < "4"
watchdog==6.0.0 ; python_version >= "3.11" and python_version < "4"
werkzeug==3.1.4 ; python_version >= "3.11" and python_version < "4.0"
wrapt==1.17.3 ; python_version >= "3.11" and python_version < "4.0"
xmltodict==1.0.2 ; python_version >= "3.11" and python_version < "4.0"
zipp==3.23.0 ; python_version >= "3.11" and python_version < "4.0"
